+++ src/cutadapt/__main__.py
@@ -823,13 +823,10 @@ def main(cmdlineargs=None, default_outfile=sys.stdout.buffer):
             runner_class = ParallelPipelineRunner
             runner_kwargs = dict(n_workers=cores, buffer_size=args.buffer_size)
         else:
-            logger.error('Running in parallel is currently not supported for '
+            parser.error('Running in parallel is currently not supported for '
                 'the given combination of command-line parameters.\nThese '
                 'options are not supported: --info-file, --rest-file, '
-                '--wildcard-file, --untrimmed-output, '
-                '--untrimmed-paired-output, --too-short-output, '
-                '--too-short-paired-output, --too-long-output, '
-                '--too-long-paired-output, --format\n'
+                '--wildcard-file, --format\n'
                 'Also, demultiplexing is not supported.\n'
                 'Omit --cores/-j to continue.')
             return  # avoid IDE warnings below